+++ common/src/sys/stats.rs
@@ -40,8 +40,8 @@ impl<'a> System<'a> for Sys {
             }
 
             if stat.exp.current() >= stat.exp.maximum() {
-                stat.exp.change_by(-stat.exp.maximum());
-                stat.exp.change_maximum_by(25.0);
+                stat.exp.change_by(-(stat.exp.maximum() as i64));
+                stat.exp.change_maximum_by(25);
                 stat.level.change_by(1);
                 stat.health.set_maximum(stat.health.maximum() + 10);
                 stat.health
